+++ tests/testthat/test_set_colors.R
@@ -11,9 +11,7 @@ tooltips <- "tooltip"
 optimize_y <- TRUE
 linewidth <- NULL
 title <- NULL
-showLabels <- NULL
 show_labels <- TRUE
-lineInterval <- NULL
 background_lines <- 11
 data <- data.frame(
   event = 1:4, start = c("2019-01-01", "2019-01-10"),
@@ -21,7 +19,7 @@ data <- data.frame(
   FARBE = c("red", "green"),
   stringsAsFactors = FALSE
 )
-data <- vistime:::validate_input(data, start, end, events, groups, tooltips, optimize_y, linewidth, title, showLabels, show_labels, lineInterval, background_lines)
+data <- vistime:::validate_input(data, start, end, events, groups, tooltips, optimize_y, linewidth, title, show_labels, background_lines)
 
 
 test_that("color columns not existing", {